backend/requirements.txt
flask
flask-cors
Pillow
opencv-python-headless
numpy
svgpathtools
requests
trimesh
shapely
